backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
Werkzeug==2.3.7 